[EI-999] [web] refactor: Update import paths for Toast component
