+++ apps/web/app/business/hooks/numerical-guidance/custom-forecast-indicator/use-selected-custom-forecast-indicator-view-model.tsx
@@ -8,7 +8,7 @@ import { convertCustomForecastIndicatorViewModel } from '../../../services/numer
 import { useEffect, useMemo } from 'react';
 import { useWorkspaceStore } from '@/app/store/stores/numerical-guidance/workspace.store';
 import { useSelectedCustomForecastIndicatorStore } from '@/app/store/stores/numerical-guidance/custom-forecast-indicator/selected-custom-forecast-indicator.store';
-import { usePending } from '@/app/ui/components/view/hooks/usePending.hook';
+import { usePending } from '@/app/utils/hooks/usePending.hook';
 import { Indicator } from '@/app/business/services/numerical-guidance/view-model/indicator-list/indicators/indicator.service';
 
 export const useSelectedCustomForecastIndicatorViewModel = () => {
