+++ apps/web/app/business/services/error-action/action-http-error.tsx
@@ -1,4 +1,4 @@
-import { Toast } from '@/app/ui/components/view/hooks/use-toast';
+import { Toast } from '@/app/utils/hooks/use-toast';
 import {
   NetworkError,
   TimeoutError,
