+++ apps/web/app/ui/components/domain/numerical-guidance/custom-forecast-indicator/custom-forecast-indicator-create-dialog-menu/custom-forecast-indicator-create-dialog-menu.tsx
@@ -2,7 +2,7 @@
 import { DIALOG_KEY } from '@/app/utils/keys/dialog-key';
 import DialogMenu from '../../../../view/molecule/dialog-menu';
 import SelectTargetIndicatorStepDialogMenu from './steps/select-target-indicator-step-dialog-menu';
-import { useStepper } from '@/app/ui/components/view/hooks/use-stepper.hook';
+import { useStepper } from '@/app/utils/hooks/use-stepper.hook';
 import SelectSourceIndicatorStepDialogMenu from './steps/select-source-indicator-step-dialog-menu';
 
 export default function CustomForecastIndicatorCreateDialogMenu() {
