+++ apps/web/app/ui/components/domain/numerical-guidance/custom-forecast-indicator/custom-forecast-indicator-edit-dialog-menu/custom-forecast-indicator-edit-dialog-menu.tsx
@@ -9,7 +9,7 @@ import SourceIndicatorSliderGroup from '../source-indicator-slider-group';
 import Button from '../../../../view/atom/button/button';
 import Pending from '../../../../view/molecule/pending';
 import { ReloadIcon } from '@radix-ui/react-icons';
-import { useDialog } from '../../../../view/hooks/use-dialog.hook';
+import { useDialog } from '../../../../../../utils/hooks/use-dialog.hook';
 import { useCustomForecastIndicatorListViewModel } from '@/app/business/hooks/numerical-guidance/custom-forecast-indicator/use-custom-forecast-indicator-list-view-model.hook';
 import ForecastTypeToggle from '../forecast-type-toggle';
 import { useCustomForecastIndicatorsValueByMetadata } from '@/app/business/hooks/numerical-guidance/custom-forecast-indicator/use-custom-forecast-indicator-value-by-metadata.hook';
