+++ apps/web/app/ui/components/domain/numerical-guidance/indicator-board-metadata/metadata-dialog-menu/metadata-dialog-menu.stories.tsx
@@ -1,7 +1,7 @@
 import type { Meta, StoryObj } from '@storybook/react';
 
 import MetadataDialogMenu from './metadata-dialog-menu';
-import { useDialog } from '../../../../view/hooks/use-dialog.hook';
+import { useDialog } from '../../../../../../utils/hooks/use-dialog.hook';
 import { DIALOG_KEY } from '@/app/utils/keys/dialog-key';
 import { useEffect } from 'react';
