+++ apps/web/app/ui/components/domain/numerical-guidance/indicator-board-metadata/metadata-list-item/metadata-list-item.tsx
@@ -2,7 +2,7 @@ import IconButton from '../../../../view/atom/icons/icon-button';
 import { DotsHorizontalIcon } from '@heroicons/react/solid';
 import { IndicatorBoardMetadata } from '@/app/business/services/numerical-guidance/view-model/indicator-board-metadata/indicator-board-metadata-view-model.service';
 import { useSelectedIndicatorBoardMetadata } from '@/app/business/hooks/numerical-guidance/indicator-board-metedata/use-selected-indicator-board-metadata-view-model.hook';
-import { useDialog } from '../../../../view/hooks/use-dialog.hook';
+import { useDialog } from '../../../../../../utils/hooks/use-dialog.hook';
 import { DIALOG_KEY } from '@/app/utils/keys/dialog-key';
 import ExpandableListItem from '../../../../view/molecule/expandable-list-item';
 import DraggableContext from '../../../../util/draggable-context';
