+++ apps/web/app/ui/components/util/swr-provider.tsx
@@ -2,7 +2,7 @@
 import { PropsWithChildren } from 'react';
 import { SWRConfig } from 'swr';
 import { onActionHttpError } from '../../../business/services/error-action/action-http-error';
-import { useToast } from '@/app/ui/components/view/hooks/use-toast';
+import { useToast } from '@/app/utils/hooks/use-toast';
 
 export const SWRProvider = ({ children }: PropsWithChildren) => {
   const { toast } = useToast();
