+++ apps/web/app/utils/hooks/use-generate-image.ts
